Performance optimization for default Backward Euler (theta_imp=1).
1. theta_imp is set as a StaticSolverConfigSlice instead of dynamic. It is not a variable which is expected to be often modified. Concrete theta_imp enables a simpler compute graph for theta_imp=1, the default setting, since explicit coefficients do not need to be calculated.

2. Also carried out a further fvm cleanup, reducing redundant function arguments, and only have redundant exposures where useful for testing and docstrings.

The reduction in computational complexity is here restricted to theta_method_matrix_equation, where the explicit matrix + vector is significantly simplified if theta_imp=0. This leads to a compilation reduction of ~10-20% and a runtime reduction of 5-10% as evaluated from 3 testcases with newton-raphson and predictor corrector. Next stage is to reduce calculations of explicit calc_coeffs.

PiperOrigin-RevId: 618220630