Significant performance increase through more jitting.
Source and current calculations were being carried out outside the stepper in the main time loop in sim, and thus uncompiled. These are now jitted, leading to a ~25-30% runtime increase in the nonlinear solver test cases, and ~factor 2 speedup in the linear and predictor_corrector cases (which have less internal stepper iterations).

PiperOrigin-RevId: 618788405