Rename `State` to `CoreProfiles`.
Throughout TORAX, we've been referring to the set of variables that are evolved by the PDE system as the "State" or "mesh state", but these names were not properly describing what they were (many things are defined on the cell grid).

This change in naming is more descriptive and also helps better align with IMAS naming conventions.

PiperOrigin-RevId: 619208526