Introduce CoreTransport dataclass to hold transport coefficients returned to the user.
We already had TransportCoeffs, so this change simply moves that dataclass and renames it. And rather than copying the data inside AuxOutput to return to the user, this change simply returns the data to the user directly.

This dataclass is somewhat inspired by IMAS's `core_transport`, though again it doesn't actually follow that schema exactly.

PiperOrigin-RevId: 619502719