Rename Sources to SourceModels.
Naming change helps make it explicit that the sources here are not the profiles themselves but rather the models/functions that compute them. In IMAS, the profiles are called "core sources", so this just helps clarify things a bit.

We don't plan to rename the TORAX SourceProfiles dataclass - this change is really just for the models.

PiperOrigin-RevId: 619855566