Set a specific commit version for `equinox`.
Version `0.11.3` has a bug in it which rears it's nasty head when we raise equinox errors in our tests. The bug is fixed at equinox's HEAD, so this change sets the requirement to HEAD at the time of this PR's creation.

Also, this changes the way the `TORAX_ERRORS_ENABLED` env variable is set in the github workflow.

PiperOrigin-RevId: 622205480