Remove `AuxOutput` and use `SourceProfiles` inside `ToraxSimState` to output complete source profiles.
Once again, the naming change is to help align a little better with IMAS's IDS. AuxOutput was really just containing a subset of the source profiles, so this change removes that unnecessary intermediate holder.

The main trickiness in this change is that it merges the implicit source profiles from the previous time step with the explicit source profiles from the current time step to give a "best guess" of the sources at time t.

PiperOrigin-RevId: 624129270