Move the transport config out of `Config`.
We move some of the config definitions out of config.py as well to start to allow modules to define things on their own, hopefully making it easier to extend TORAX in the future (not everything needs to live in config or config_slice.py).

Moves us closer to the end state of configuration which we want.

PiperOrigin-RevId: 627850803