Add `build_sim.py`, entrypoint for building the `Sim` object from a lightweight python dict config.
Most examples we have now in TORAX are configured with a few python functions and building actual objects. We want to support a more bare-bones version of the config which uses only python primitives and dicts so people can rapidly iterate without dealing with Python objects.

This PR doesn't actually implement everything, but just lays out the structure and the docs.

PiperOrigin-RevId: 629042971