Add `list_files`.
This lists all the files in Torax. In subsequent PRs, this will be used
to compute a hash of all the Python code in Torax, so that queries to
the Jax persistent cache will correctly miss if the Torax code has been
updated (and could cause a different Jax trace) since the cache entry
was created.