Rename CHEASEGeometry to StandardGeometry.
We are going to start supporting geometries coming from different sources, not just CHEASE, and many of them will have the same fields as the current CHEASEGeometry class.

As a precursor before those changes come in, we want to rename this class and have it be a bit more generic.

PiperOrigin-RevId: 642646652