Add QLKNN model path to the config and move flag to main.
The order of priority to set the model path are now: flag, config, env_var, default hard-coded path.

This allows users to set the model path in the config, which can be convenient when running inside another process for instance.

This CL also changes the way that the QLKNN model is cached. We get rid of global variables. The loading uses an lru_cache dependent on the model_path.

The hash and equality methods of the QLKNNTransportModel is now using the model_path, so we can truly know whether two QLKNNTransportModel objects are the same.

PiperOrigin-RevId: 644383277