Ignore extra kwargs in stepper runtime params.
Presently all stepper runtime params are flat in config['stepper']. This leads to issues such as raising KeyError when switching a stepper from NewtonRaphson to Linear, and keeping log_iterations in the config. This is not user-friendly since it is beneficial to toggle steppers for testing/comparison and not need to worry about spurious fields from other stepper runtime params.

PiperOrigin-RevId: 647396511