Add Phibdot to geometry dataclasses.
Phibdot is a special case since it is a time derivative and depends on having both geo_t and geo_t_plus_dt provided. Following their provision Phibdot is calculated and both geo dataclasses corresponding to the time interval are updated with this same value.

Phibdot will be used in upcoming PRs in calc_coeffs to calcuate terms related to time-dependent geometry.

PiperOrigin-RevId: 655093793