Refactor of terms and remove geo.J.
Refactors for consistency between the psi equation terms in calc_coeffs and source_models.calc_psidot.

Also removing the usage of geo.J in the equations, to align with the published form of the governing equations.

A step towards various equation cleanups, including modifying the equations to the published rhonorm form, which is an upcoming PR

PiperOrigin-RevId: 656027980