I had either misunderstood the Jax requirements and/or the Jax requirements have changed.
It is true that all of our arguments to Jax functions need to be immutable
and hashable in order for the Jax in-memory cache to work correctly.
Mutable Sources were probably the cause of the run_simulation_main cc
bug (old runtime_params were compiled into the Jax program and new runtime
params didn't cause a cache miss)

The part I had wrong is the description that the *persistent* cache maps
from lower arg keys to the compiled program and thus skips tracing.
Actually tracing is still done when we re-run the python script, the
persistent cache key is a function of the built graph, not the args to
function where the jit decorator is applied. This means that our hash
functions for custom classes that are arguments to the outermost function
don't need to be designed to hash the same across runs of the Python
interpreter.

Some of the hash functions I had already written were failing at that
anyway because they were based on the built in `hash` function. In newer
versions of python3 the `hash` function gets a different random seed
each time the interpreter is started. This is what led me to notice that
Jax doesn't actually need stable keys from us.

Because these features
are complicated, easy to get wrong, and not needed, I'm removing them.
It's a subtle point though---I'm getting rid of *persistent* hashing,
but we still need to maintain immutability, and hashability within
a single Python process.

PiperOrigin-RevId: 656084743