Fix update_psi_from_j formula.
Previously we were calculating currents with dV/drhon instead of dS/drhon (volume instead of area). This is incorrect since j [A/m^2].

This error was compensated by a 2*pi*R in the scale factor for dpsi/drhon, which was an exact correction for circular geometry, but in the data coming from CHEASE geometry led to slight errors on order of 1% for psi. The two CHEASE psi_from_j reference cases were recalculated.

This PR also renames a couple of related variables for clarity.

PiperOrigin-RevId: 658572435