Fix jext source to be time-dependent and recalculate on each timestep.
This was a bug where we only calculated the jext profile once at the start of the sim run, and used the same value every time afterwards. Now the profile is computed on every step.

This also changes the source to only output the face profile, and we compute the cell profile when needed, thus simplifying the code a little bit.

PiperOrigin-RevId: 660379223