Add error handling for NaNs in core profiles.
Adds a new `SimError` enum to track simulation errors. The `sim.run()` function now returns a tuple containing the simulation outputs and the `SimError`. If the simulation encounters an error, it will stop and return the error code.

All profiles up to the last valid timeslice are still saved to file as an output xarray dataset.

The `simulation_app.py` script has been updated to record the `SimError` in the output file. This may be useful e.g. when carrying out large-scale parameter scans, to identify non-physical / failing cases.

A new test case has been added to test the error handling. The test case sets up a simulation that evolves to negative densities via a fictitious negative density source, which causes NaNs to appear in the core profiles. The test case verifies that the simulation stops early and returns the correct error code.

The test_run_simulation_main nc files were updated since the sim_error attribute is also compared.

PiperOrigin-RevId: 673014636