Add support for Zeff to be a non-constant radial profile.
Also support a single float to be allowed input for InterpolatedVarTimeRho, being a short cut for constant over time and space. This is useful regardless, and also maintains compatibility with the existing Zeff default and test config files.

Add new test based on the iterhybrid_predictor_corrector, with non-constant Zeff.

PiperOrigin-RevId: 677722208