Obtain vacuum B, Rmajor, Rminor directly from FBT geometries.
Previously we were inputting them separately. This was superfluous, this information is in the FBT data itself. Removing the superfluous data input now enables inputting a bundled LY file with all time slices therein, e.g. as created by liuqe meqlpack. This feature will come in an upcoming PR.

Also fixed an unnecessary (was near unity) rescaling of Phi in from_fbt. LY['FtPQ'] is already the total toroidal flux, both vacuum and plasma part, which is what we need.

Differences in FBT simulations are minor, on the 1e-3 level.

PiperOrigin-RevId: 687456705