Modify run_simulation_main test to be robust to validity check.
At the moment this test fails when run with pytest on a single process (see discussion on https://github.com/google-deepmind/torax/pull/443). I think this is because the handlers between tests interact in a strange way and some tear down might not be being done correctly.

Instead of digging into it further the modification of checking the output file is saved and equals a reference seems a more robust check that the simulation "completed" than checking the captured logged output and passes on a single process with pytest.

PiperOrigin-RevId: 691465286