Fix stitch feature in sim restarts.
Previously, the drop_duplicates was causing unexpected behavior when previous_ds contained many of the same times as ds. It is cleaner to just cut previous_ds to a smaller slice and just drop the single duplicate.

The restart sim test is also extended to check all ds outputs. This exposed a bug where sigma (collisionality) was not updated properly. This was fixed as a drive-by.

All test simulations were rerun to incorporate the correct sigma outputs.

PiperOrigin-RevId: 691749025