Fix bug in Coulomb constant.
Density was normalized wrongly by an order of magnitude.

Coulomb constant however is relatively insensitive, and sim test differences are only different by O(1e-3). Sim tests generated.

Acknowledgements to Theo Brown for spotting this.

As a drive-by, this PR exposed a bug in the restart logic, uncovered due to a newly different adaptive timestep in test_iterhybrid_rampup.

PiperOrigin-RevId: 692750143