Clean up physics.
Reduce duplication by defining functions for repeated calculations.

coll_exchange and QLK were not using consistent definitions of ion-electron collisional times. The one in QLK was replaced by the more exact version. Differences to results are small on the O(1e-3)-O(1e-5) level, with the exception of test_iterhybrid_newton which is on the O(1e-2) relative level, due to a different trajectory taken.

PiperOrigin-RevId: 692899879