Fix EQDSK geometry coupling and add a new ITER hybrid test case.
New ITERhybrid EQDSK file generated from CHEASE data. Good comparison with the
simulation using CHEASE geometry.

Summary of fixes:
1. vpr definition bug leading to major impact on plasma volume
2. Avoiding LCFS flux-surface-integral issues for diverted cases by taking the last contour just inside the LCFS.
3. Rmaj taken now as Rgeo (LCFS Rmaj), and local Rmaj and Rmin used for delta (triangularity) calculations
4. StandardGeometryIntermediate values at the magnetic axis are prescribed, since a contour cannot be defined there.

The number of surfaces to calculate, and the location of the LCFS contour with respect to the boundary psi, are user-inputs, with reasonable defaults set but need to be checked against more varied sets of EQDSK files.

The EQDSK converter has only been tested against CHEASE-generated EQDSK which is at a specific COCOS. Therefore issues may arise if the input EQDSK is in a different COCOS. In near-future work we should demand that the EQDSK COCOS is an input, and we use the eqdsk convertor to convert to the TORAX COCOS, which itself also needs to be made fully self-consistent in a future PR.

This important caveat is mentioned in the documentation, as well as a user-warning when running TORAX with EQDSK geometry.

PiperOrigin-RevId: 697621831