Upgrade to XArray DataTree.
Currently this is only a partial upgrade to test persistence and loading of DataTree and general usage. It doesn't change much in terms of usage of Torax outputs as it uses the top level dataset that can be stored in the DataTree object to hold the previous dataset.

Follow up change to use hierarchy in the DataTree and change how the output users use it.

PiperOrigin-RevId: 697988945