Expand documentation on persistent cache.
A useful thing to know about the persistent cache is that
the hash of an object to be included in the key only needs to be constant across a given run and not between runs as helpfully explained in https://github.com/google-deepmind/torax/pull/276.

Also update an outdated comment that says otherwise.

PiperOrigin-RevId: 698339770