TEMP: Initial attempts at getting Vloop BC working
This includes a hack to support outputting different BCs in the output
file. Previously, the simulation could only use one type of BC for the
whole run (ie either grad or value constraint). By wrapping the output in
a jnp.array(), BCs that are None get turned into NaN, which is compatible
with tree_map. Hence, this change allows you to have `grad_constraint =
[XXX, None, None, ...]` and `value_constraint = [None, XXX, YYY, ...]`
which is useful for testing the Vloop BC.