Divide up source runtime params into static and dynamic attributes.
Fields like `mode` and `is_explicit` don't need to be in dynamic as we don't mind triggering a recompile on them. This change moves those two fields into a newly created `StaticRuntimeParams` for sources and then adds this object into the `StaticRuntimeParamsSlice`. A hash needs to be generated for the new `StaticRuntimeParamsSlice` as the autohash doesn't work now a `Mapping` has been introduced.

This also opens up a path for specifying more flexible modes within static params than needing to use one of the four predefined modes now.

PiperOrigin-RevId: 702367926