Set the mesh for interpolated params in `ToraxConfig`.
This also generalizes `TimeVaryingArray` to accept both `grid_face_centers` and `grid_cell_centers`. The `get_value` method is generalized to accept an option of which to use for the interpolation. Given that `get_value` is not different from the `TimeVaryingScalar` case and there are now two cached_properties, there is little value in having an interpolated param base class. `set_rho_norm_grid` is now also removed.

PiperOrigin-RevId: 733256132