Integration routines in math_utils to simplify calculations.
Also replace various integrations throughout TORAX with the math_utils versions.

This includes the source formula integrations, which were previously done on the face grid using the trapezoid rule when generating the face values using the formulas (as opposed to integration conserving math_utils.cell_to_face). This means that the integrals are now slightly different. The present version is more consistent since it truly treats the cell values as cell averages, and uses the appropriate volume integration.

Sim benchmarks are regenerated. Relative differences are on the O(1e-3) level.

PiperOrigin-RevId: 734195350