Remove `from __future__ import annotations`.
This is largely incompatible with runtime type checking (and removing unlocks the option to use jaxtyping for runtime checking).
Prefer to use stringified types directly where needed and avoid them if not needed.

Feel free to push back on this CL if you'd rather keep the `from __future__ import annotations` pattern.

PiperOrigin-RevId: 737595931