Allow annotating which Pydantic model fields are considered static by JAX (jax.jit, jax.tree.flatten, etc), so that the model can be used directly as a JITted function argument.
A design question: some magic is possible here (eg. strings will always be JAX static args. Could avoid needing to annotate them). For the first version: decided to follow the JAX design and require explicit annotations.

This is required for followup caching functionality.

This provides an alternative to the Torax dynamic params mechanism, as manual splitting of params into dynamic and static is no longer required.

PiperOrigin-RevId: 737978420