Make face_centers and cell_centers properties on Grid1D.
Critical here is the use of JAX_STATIC annotation: without this, nx and dx can be JAX tracers in JIT, which makes computing linspace impossible (linspace requires looking at the actual value of nx), and disallows global caching for efficiency (JAX tracers are not a hashable type).

Significantly reduces size of serialized model.

PiperOrigin-RevId: 738041927