Enable jax_utils.jit to be reconfigured at runtime rather than import time.
This is much more expected behaviour (and is used assuming this around the code atm). Requires care in unsetting the environment variable as well but overall worth it imo.

This also allows the tests such as sim_no_compile to actually run without compilation which are at the moment silently running with compilation.
Also exposed the optimizer test is broken when running with JIT compile turned off atm. Removed this from the sim_no_compile tests.

PiperOrigin-RevId: 740482734