Fix stepper discriminator logic.
The callable discriminator did not handle the serialization case where the model can be a pydantic model and not a dict, and thus would produce warnings (`Failed to get discriminator value for tagged union serialization`). This removes the callable discriminator logic and replaces it with simpler string discriminator logic that cannot suffer from this issue.

PiperOrigin-RevId: 742373220