More streamlined installation and use of QLKNN_7_11 transport model
This CL adds the `fusion_surrogates` package to the pip config. This will ensure that `fusion_surrogates` is always available.

We also update the model loading to remove the requirement for a model path. `fusion_surrogates` now has a registry of models, so we can load them by name, or just load the default model.

These two changes removes all manual steps that used to be required to setup the transport model upon TORAX installation.

The new logic for choosing a model to load is:
 - look at model_path in the config
 - if not defined, look at the env variable
 - if not defined, look at model name in the config
 - if not defined, load the default model from `fusion_surrogates`

I also updated the doc to reflect this, as well as updating some deprecated documentation on transport configs.

PiperOrigin-RevId: 745284106