Prevent two subsequent sawtooth crashes in a row.
This change protects against a sawtooth crash if it occurs twice in a row. This can happen if the redistribution in insufficient to bring q=1 below the critical shear. In this case it may happen that standard PDE steps are never carried out from that point onwards.

This PR also introduces the "sawtooth_just_crashed" attribute to ToraxSimState, which is True if the state corresponds to post-sawtooth-crash. Beyond being used to avoid consecutive crashes, this attribute is useful for simulation diagnostics.

Reran all the sim tests to obtain the new attribute.

Drive-by: extended the general sawtooth integration tests when writing the specific test for this CL.
PiperOrigin-RevId: 745339451