Remove __hash__ and __eq__ from `CellVariable` and __hash__ from `CoreProfiles`.
As far as I understand there is no reason for `CoreProfiles` to be hashable and the equality check was being done on `id` which was also causing the file restart test to think two objects weren't equal because they didn't have the same id when I believe they should be checking the trees themselves aren't equal.

Removing __eq__ will stop this easy mistake in future.

PiperOrigin-RevId: 747803121