Remove all face variables from core profile outputs.
Users have comment it is confusing for TORAX outputs being defined sometimes on multiple grids or in multiple variables e.g.
- jbootstrap is a combination of cell and face grids in jbootstrap and j_bootstrap_face
- temp_el (and other profiles) have a separate boundary value entry

Additionally some of the face values are interpolated quantities from the cell values. Again it's additional overhead for users to know which quantities are interpolated and which are the "true" values.

This change is the first in consolidating all of these into a single "cell plus boundaries" grid so that there are simply computed "profile" quantities with grids defined by the coordinates.

This moves relevant variables over to the cell plus boundaries grid.

PiperOrigin-RevId: 750961471