Add Z_eff formula and output variable.
Currently Z_eff is only provided in input PlasmaComposition. It is useful to have it also as an output. Sim tests regenerated with new Z_eff output. Sim tests themselves were passing before the update.

Furthermore, in future TORAX extensions Z_eff may be a predicted quantity and not an input quantity. Therefore we should shift to CoreProfiles as the source of truth for Z_eff.

A realization was made in the course of this PR, that using plasma_composition_Z_eff_face in physics calculations is incorrect, since that does not correspond to the actual Z_eff_face according to n_i.face_value() and n_impurity.face_value() due to a lack of linearity between the dilution calculation and Z_eff. A followup PR will change the SoT of Z_eff to CoreProfiles in internal calculations. This is expected to lead to small modifications of simulations with radially varying Z_eff

PiperOrigin-RevId: 767153252