Add utilities for preventing compiler inlining of JAX functions.
This is useful for reducing compile times.

Two methods are given as they have different pros and cons: the pure_callback method allows for true reuse of compiled blocks (if the same function is called multiple times in an XLA program, it will be compiled only once). The while_loop method has much less overhead, but will still compile multiple times if the same function is used. However: it allows for a linearization of compile-times (XLA compile time increases extra-linearly with increasing ops otherwise).

PiperOrigin-RevId: 773301267