Add neoclassical models to static args of experimental compile of newton raphson and run experimental compile and sim no compile tests as intended.
Copied over the template from `sim_no_compile_test` without thinking too much and that test had not been running as intended (without compilation on), this also meant a bug got introduced on syncing to HEAD with the new neoclassical models that was uncaught (as the experimental compile pathway was not being hit).

Fixed bug and tests.

PiperOrigin-RevId: 773731991