Move error check outside `StepFunction`.
This change is a design question. At the moment the `StepFunction` takes in a `ToraxSimState` and `PostProcessedOutputs`. For symmetry it feels clean for it also to return the same.

I'd also like to open up the discussion on whether post_processed_outputs should be returned from this method or whether it should simply take in and return the `ToraxSimState` and the post processed outputs be dealt with separately.

This also gives the StepFunction a single responsibility of stepping and error handling will be done by something else.
This will be useful when JITting the run loop as well as error handling may need to be treated differently under JIT.

PiperOrigin-RevId: 775131092