Add dataclass decorator for dataclasses registered with JAX.
This behaves like a regular dataclass. Note that chex.dataclasses also have support for behaving like a `Mapping`. Is this functionality something we want? In TORAX atm we do make use of this but not really in a way that makes me think we need it but keen to hear opinions!

PiperOrigin-RevId: 776050947