New pattern for default values for discriminated pydantic model configs.
Avoids need for before validator logic in each pydantic_model for default model setting. The logic is always done by the new function in torax_pydantic. The responsibility of the user is just to provide the discriminator name and default model name when configuring the new before validator for each pydantic model.

Demonstrated on neoclassical transport.

Followup PR can implement this in all other models.

PiperOrigin-RevId: 779589294