Mark `ConstantGeometryProvider` as a Pytree.
This completes marking all `GeometryProvider`s as a Pytree.

Also added __eq__ method to all attribute types on the `GeometryProvider`s. This was to allow the ToraxConfig (de)serialisation test to pass correctly as it was previously only checking equality on id and switching to dataclass made it compare on attributes.

PiperOrigin-RevId: 783892636