Correct calculation of average Z for a mixture of ions.
Changes the calculation of the average Z from a straight-average to a Zeff and main-ion-diluting respecting average.

This means that the effective impurity charge is now corrected, as well as the effective Z for the main ion if different Z are used.

More details are included in the function docstring.

New getter tests are written, comparing Zeff and main-ion-dilution expectations to what we would expect if directly setting impurities as n_e fractions, to what is obtained with fractional distribution inputs and imposed Z_eff, and the new effective charge formulas.

mavrin_impurity_radiation sim test is regenerated since that one had multiple impurities and thus a new Z_impurity with the new formulation. Relative differences in Ti, Te, ne, psi only on order of O(1e-3)

PiperOrigin-RevId: 784518585