Add impurity fractions to CoreProfiles
i) Storing the fractions in CoreProfiles following avoids doing duplicate fraction calculations when implementing the upcoming 'n_e_ratios_Z_eff' impurity_mode. For example, in the mavrin radiation model, fractions are needed. It's best if they are stored after various ion calculations in updaters.py, and then reused downstream where needed instead of recalculating fractions which involves charge state calculations, etc.

ii) 'n_e_ratios_Z_eff' impurity mode can possibly introduce a negative individual impurity density. e.g, if input Z_eff is too low compared to the input n_e_ratios for the other impurities. The bundled n_impurity can be positive, but the individual fraction negative. This is unphysical and needs to be caught by the sim state error checking.

iii) Useful information on the fractional abundances of impurities in the output structures. Currently this information is only in the input config. However, pending discussion on the final new API for impurities, we suppress this output to the public API (the DataTree). For example, we may want to proceed to directly output the individual impurity densities directly, and not just the fractions.

PiperOrigin-RevId: 789882251