Remove formatting workflow.
Autoformatting done for TORAX internally to Google has a bug in it that means our formatting is less strict than the external pyink and isort checks, as such changes we make keep failing the external formatting check which is producing some friction. As all external changes get passed through our internal formatter anyway, propose removing the external formatting workflow for now till the internal autoformat bug is fixed. Developers are still able and encouraged to use the `pyink` and `isort` on created PRs.

PiperOrigin-RevId: 794511695