Refactor to specify impurity_mode as 'fractions'.
Part of series of PRs to allow specifying the impurity type as either fractions, or n_e ratios with optional Z_eff constraint.

The impurity type is handled by discriminated unions in pydantic.

This PR solely implements the changes for the 'fractions' option which is the present option, and is thus a refactor

This is implemented with backward compatibility for the V1 API. A `model_validator` method on PlasmaComposition handles this conversion for backward compatibility, logging warnings for deprecated behaviour.

PiperOrigin-RevId: 794519576