Add public script to regenerate TORAX reference values.
This CL introduces `regenerate_torax_refs.py`, a script for recalculating and updating the reference values used in TORAX unit tests. It defines a `REFERENCES_REGISTRY` in `torax_refs.py` to map case names to functions generating the reference configurations.

The script can regenerate all or specific cases, print detailed summaries, and save the updated values to a `references.json` file.

This replaces a colab that previously was not publicly available.

The next PR will implement the json file, modify torax_refs to read from the json and remove the data in torax_refs.

PiperOrigin-RevId: 798171065