Modify ImpurityFractions to contain a `TimeVaryingArray` of fractions.
Also update getters.py logic to handle fractions defined on the rho grid rather than constant radial fractions.

Create a separate module for impurity fractions with its own `RuntimeParams` to clearly separate responsibility from `IonMixture`.

A "legacy" flag is added to inputs that are flagged as part of the V1 API in `PlasmaComposition`. This flag is used to signal to the `ImpurityFractions` to convert inputs from legacy `TimeVaryingScalar`s inputs to equivalently functioning `TimeVaryingArray`s.

PiperOrigin-RevId: 805892108