Introduce PedestalPolicy.
This is a step toward having full Martin scaling, where there will be a
Martin scaling PedestalPolicy that ramps up and ramps down the pedestal
height. This PR is just a step that defines the new stateful interface
required for doing so and transitions the existing `set_pedestal`
functionality to being implemented with the new PedestalPolicy class.

PiperOrigin-RevId: 810477225