Correct Phi_b_dot calculation and set at object creation time instead of post-populating.
Sets the calculation of Phi_b_dot to the "true" gradient as the existing one is inaccurate if Phi_b is specified to vary within the time interval.

Also move the calcphibdot term from numerics to geometry and add logic to ToraxConfig to parse v1 API numerics.calcphibdot field into geometry if present.

This sets up simplifying the step function logic in a follow up.

PiperOrigin-RevId: 816655867