Clip c_z_prefactor to zero in inverse mode lengyel and add test
It turned out that q_div_squared < 0.0 is a difficult condition to meet, and a more intuitive physics consistency check is directly on c_z_prefactor. PhysicsOutcome (previously SolveStatus) enum changed to reflect this. c_z_prefactor clipped to zero in analogy to q_cc in forward mode.

Later, the PhysicsOutcome enum will be propagated to extended-lengyel outputs as diagnostic metadata.

Drive-by: Rename SolveStatus to PhysicsOutcome which is a more accurate description. Later we will add another enum for the actual solver status.
PiperOrigin-RevId: 819343944