MHDModels cleanup
MHDModels was registered as a pytree, so it looked like it's meant to be
a dynamic argument.
But it also has an __eq__ and __hash__ method, so it looks like it's meant
to be a static argument.
Besides that, the __eq__ and __hash__ methods aren't safe and it wouldn't
work as a static argument because it wasn't actually frozen.

I confirmed that it is used as a static argument: it is a field of
PhysicsModel registered as static.

This CL fixes the problem where it was not frozen, adds a comment that it
is intended for use as a static argument, and removes the unnecessary
pytree registration.

Also, now that it is a frozen dataclass, the __hash__ and __eq__ methods
may be removed, because the frozen dataclass defaults work. (This class
has no polymorphisms and doesn't need to hash the class id). The existing
__eq__ method had a bug where it assumed `other` is an MHDModels.

PiperOrigin-RevId: 819877973