Fix bugs in Solver hash and eq.
Solver is used a static argument. Solver is polymorphic and the subclass identity is used as an implicit argument affecting the trace, so the subclass identity must be part of the key for the Jax cache. The Solver __hash__ and __eq__ methods are not hashing and comparing the class id. This CL fixes that by comparing a string representation of the class id, including the module path.

PiperOrigin-RevId: 819912183