Add edge-relevant geometry parameters to `StandardGeometry`.
This change introduces optional geometry parameters related to the edge, such as connection lengths and target angle, to the `StandardGeometry` dataclass. The `TimeVaryingGeometryProvider` is updated to handle the interpolation of these new parameters, including special handling for the boolean `diverted` flag.

Implementing the edge geometry parameters in the various geometry type constructors will be done on a case-by-case basis when ready, starting with FBT/MEQ.

PiperOrigin-RevId: 825570190